Rename pkg/async to pkg/asyncworker
